add ocr and ngrok api key to test ocr on smartphone

+++ backend/src/main/resources/db/migration/V1__create_schema.sql
@@ -1,18 +1,18 @@
 CREATE TABLE brand (
-    id      SERIAL PRIMARY KEY,
+    id      BIGSERIAL PRIMARY KEY,
     name    VARCHAR(50)  NOT NULL,
     country VARCHAR(50)
 );
 
 CREATE TABLE vehicle_range (
-    id       SERIAL PRIMARY KEY,
-    brand_id INTEGER     NOT NULL REFERENCES brand(id),
+    id       BIGSERIAL PRIMARY KEY,
+    brand_id BIGINT      NOT NULL REFERENCES brand(id),
     name     VARCHAR(50) NOT NULL
 );
 
 CREATE TABLE vehicle_model (
-    id          SERIAL PRIMARY KEY,
-    range_id    INTEGER      NOT NULL REFERENCES vehicle_range(id),
+    id          BIGSERIAL PRIMARY KEY,
+    range_id    BIGINT       NOT NULL REFERENCES vehicle_range(id),
     name        VARCHAR(100) NOT NULL,
     year_from   SMALLINT,
     year_to     SMALLINT,
@@ -21,8 +21,8 @@ CREATE TABLE vehicle_model (
 );
 
 CREATE TABLE vehicle_specs (
-    id                  SERIAL PRIMARY KEY,
-    model_id            INTEGER      NOT NULL UNIQUE REFERENCES vehicle_model(id),
+    id                  BIGSERIAL PRIMARY KEY,
+    model_id            BIGINT       NOT NULL UNIQUE REFERENCES vehicle_model(id),
     engine_description  VARCHAR(100),
     fuel                VARCHAR(20),
     displacement_cc     INTEGER,
@@ -38,7 +38,7 @@ CREATE TABLE vehicle_specs (
 );
 
 CREATE TABLE plate_lookup (
-    id       SERIAL PRIMARY KEY,
+    id       BIGSERIAL PRIMARY KEY,
     plate    VARCHAR(10) NOT NULL UNIQUE,
-    model_id INTEGER     NOT NULL REFERENCES vehicle_model(id)
+    model_id BIGINT      NOT NULL REFERENCES vehicle_model(id)
 );
